backend/src/requirements.txt
flask==3.0.*
flask-cors==4.0.*
elasticsearch==8.13.*      # same client for local & OpenSearch
python-dotenv==1.0.*
boto3==1.34.*               # Lambda in VPC needs it anyway
eventregistry==9.1
requests
opensearch-py
requests-aws4auth
